Develop

+++ app/Http/Controllers/PostController.php
@@ -6,8 +6,8 @@
 
     use App\Http\Requests\PostRequest;
     use App\Models\Post;
-    use Illuminate\Http\Request;
     use Illuminate\Http\Response;
+    use Spatie\Permission\Exceptions\UnauthorizedException;
 
 
     class PostController extends Controller
@@ -19,6 +19,9 @@
          */
         public function index()
         {
+            if (!auth()->user()->hasPermissionTo('Consult')) {
+                throw new UnauthorizedException('403', 'Custom Message');
+            }
             $posts = Post::all();
 
             return view('posts.index', [
@@ -33,6 +36,10 @@
          */
         public function create()
         {
+            if (!auth()->user()->hasPermissionTo('Register')) {
+                throw new UnauthorizedException('403','Custom Message');
+            }
+
             return view('posts.create');
         }
 
@@ -44,6 +51,10 @@
          */
         public function store(PostRequest $request)
         {
+            if (!auth()->user()->hasPermissionTo('Register')) {
+                throw new UnauthorizedException('403','Custom Message');
+            }
+
             $post = new Post();
             $post->title = $request->title;
             $post->text = $request->text;
@@ -78,6 +89,9 @@
          */
         public function edit(Post $post)
         {
+            if (!auth()->user()->hasPermissionTo('Edit')) {
+                throw new UnauthorizedException('403', 'Custom Message');
+            }
             return view('posts.edit', [
                 'post' => $post
             ]);
@@ -92,6 +106,9 @@
          */
         public function update(PostRequest $request, Post $post)
         {
+            if (!auth()->user()->hasPermissionTo('Edit')) {
+                throw new UnauthorizedException('403', 'Custom Message');
+            }
             $post->title = $request->title;
             $post->text = $request->text;
 
@@ -113,6 +130,9 @@
          */
         public function destroy(Post $post)
         {
+            if (!auth()->user()->hasPermissionTo('Delete')) {
+                throw new UnauthorizedException('403', 'Custom Message');
+            }
             $post->delete();
             return redirect()->route('post.index');
         }
